+++ assets/js/base/hooks/index.js
@@ -7,6 +7,7 @@ export * from './use-collection';
 export * from './use-collection-header';
 export * from './use-collection-data';
 export * from './use-container-queries';
+export * from './use-local-storage-state';
 export * from './use-position-relative-to-viewport';
 export * from './use-previous';
 export * from './use-shallow-equal';
